+++ src/view/sorting.js
@@ -1,6 +1,7 @@
 import AbstractView from './abstract.js';
+import { SortType } from '../const.js';
 
-export default class SortingView extends AbstractView{
+export default class SortingView extends AbstractView {
   constructor() {
     super();
 
@@ -8,7 +9,6 @@ export default class SortingView extends AbstractView{
   }
 
   _sortTypeChangeHandler(evt) {
-
     if (evt.target.tagName !== 'LABEL') {
       return;
     }
@@ -26,27 +26,27 @@ export default class SortingView extends AbstractView{
   getTemplate () {
     return `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
       <div class="trip-sort__item  trip-sort__item--day">
-      <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day">
+      <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="${ SortType.DAY }">
       <label class="trip-sort__btn" for="sort-day">Day</label>
       </div>
 
       <div class="trip-sort__item  trip-sort__item--event">
-      <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event">
+      <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="${ SortType.EVENT }">
       <label class="trip-sort__btn" for="sort-event">Event</label>
       </div>
 
       <div class="trip-sort__item  trip-sort__item--time">
-      <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time">
+      <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="${ SortType.TIME }">
       <label class="trip-sort__btn" for="sort-time">Time</label>
       </div>
 
       <div class="trip-sort__item  trip-sort__item--price">
-      <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price" checked>
+      <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="${ SortType.PRICE }" checked>
       <label class="trip-sort__btn" for="sort-price">Price</label>
       </div>
 
       <div class="trip-sort__item  trip-sort__item--offer">
-      <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
+      <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="${ SortType.OFFER }" disabled>
       <label class="trip-sort__btn" for="sort-offer">Offers</label>
       </div>
     </form>`;
